agenda update: type checking
needed to implement nth, wich branches to ia-nth or ba-nth depending
on the input type.  note that in order for this to work, heap and
inline-code representations should look the same.  this entails
changing the bastart instrution to also skip over a header and leave
the pointer on the head of the data structure.