support application-defined error handler
If the application defines a word `on-error` in any dictionary,
in case of error, that word is called, with error information
on the stack, along with any items that have been on the stack
when the error happened.  Calling `listener` in that custom handler
gives a repl to debug stuff.