add task: `compile_all`
When called, outputs a combined control/data flow graph for each
definition that is in the image.  Performs stack depth checking,
inlining.  Does not compile combinators, although for declared quotation
stack effects that should be doable, even when no code can be generated
for the combinator only in a language that has no higher order functions.