remove remnants of memory bounds checking
There used to be built_in support to make sure that memory access
primitives can only access sections provided by the linker.  That
proved to be not useful, so the old code can go away now.

This changes the number of items provided by the `memrange` primitive to
exclude the two obsolete values, which were not used anywhere anyways.